[Wayland] Partial implementation of DPI handling.
The correct value of `config.dpi` is required to support physical units
in the themes (e.g. `mm` in sidebar-v2).
While we can't calculate the value without knowing the output for a
surface, we can copy the branches that don't require auto-detection. And
one of these covers the default `config.dpi` value of `-1`.